🔨 Refactor native android tests
